[Refactor] Simplify B12X linear kernels and warmup (#52368)
