+++ .buildkite/test_areas/kernels.yaml
@@ -339,6 +339,28 @@ steps:
     # e2e
     - pytest -v -s tests/models/quantization/test_nvfp4.py
 
+- label: B12X Linear Kernels (DGX Spark) Nightly
+  key: b12x-linear-kernels-dgx-spark-nightly
+  timeout_in_minutes: 30
+  device: dgx-spark
+  optional: true
+  num_devices: 1
+  depends_on:
+  - arm64-image-build
+  source_file_dependencies:
+  - setup.py
+  - vllm/model_executor/kernels/linear/
+  - vllm/model_executor/warmup/b12x_warmup.py
+  - vllm/utils/b12x.py
+  - tests/model_executor/kernels/test_b12x_linear.py
+  - tests/model_executor/test_b12x_warmup.py
+  - tests/kernels/quantization/test_block_fp8.py
+  commands:
+  - uv pip install --system b12x==1.2.4
+  - pytest -v -s model_executor/kernels/test_b12x_linear.py
+      model_executor/test_b12x_warmup.py
+      kernels/quantization/test_block_fp8.py -k b12x
+
 - label: Kernels Helion Test
   key: kernels-helion-test
   timeout_in_minutes: 115
